Separate RESTful actions methods into their own modules
This is a preparation for the restful_actions DSL, that should only
define methods related to the actions described in the resource.

Also, if you think this through, it doesn't make much sense to keep these
methods under a "Collection" module. `find` for example is not a
collection method. As it always returns one element, it's a member.